# lib/api-zod/src/generated/types/aiResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AiResponse {
  result: string;
  action: string;
}