# lib/api-zod/src/generated/types/highlight.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Highlight {
  id: number;
  bookId: number;
  userId: number;
  selectedText: string;
  /** @nullable */
  pageNumber?: number | null;
  color: string;
  createdAt: string;
}